+++ src/Controller/Customer/Card/Response/Success.php
@@ -11,8 +11,24 @@
 
 namespace Verifone\Payment\Controller\Customer\Card\Response;
 
+use Magento\Framework\App\Action\Context;
+use Magento\Framework\App\Request\Http;
+
 class Success extends \Magento\Framework\App\Action\Action
 {
+
+    public function __construct(Context $context)
+    {
+        parent::__construct($context);
+
+        // Fix for Magento 2.3 CsrfValidator and backwards-compatibility to prior Magento 2 versions
+        if(interface_exists("\Magento\Framework\App\CsrfAwareActionInterface")) {
+            $request = $this->getRequest();
+            if ($request instanceof Http && $request->isPost()) {
+                $request->setParam('ajax', true);
+            }
+        }
+    }
 
     public function execute()
     {